requirements.txt
# Core
python-dotenv>=1.0
# supabase REST via httpx (no supabase-py needed)
websockets>=12.0

# Data / ML
numpy>=1.26
pandas>=2.2
xgboost>=2.0
scikit-learn>=1.4

# Dashboard
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1

# HTTP
httpx>=0.27
